@@ -12,52 +12,55 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^5.2.7",
-    "@angular/cdk": "^5.2.4",
-    "@angular/common": "^5.2.0",
-    "@angular/compiler": "^5.2.0",
-    "@angular/core": "^5.2.0",
-    "@angular/forms": "^5.2.0",
-    "@angular/http": "^5.2.0",
-    "@angular/material": "^5.2.4",
-    "@angular/platform-browser": "^5.2.0",
-    "@angular/platform-browser-dynamic": "^5.2.0",
-    "@angular/router": "^5.2.0",
-    "angular2-image-upload": "^1.0.0-rc.0",
+    "@angular/animations": "^6.0.5",
+    "@angular/cdk": "^6.2.1",
+    "@angular/common": "^6.0.5",
+    "@angular/compiler": "^6.0.5",
+    "@angular/core": "^6.0.5",
+    "@angular/forms": "^6.0.5",
+    "@angular/http": "^6.0.5",
+    "@angular/material": "^6.2.1",
+    "@angular/platform-browser": "^6.0.5",
+    "@angular/platform-browser-dynamic": "^6.0.5",
+    "@angular/router": "^6.0.5",
+    "@ngx-pwa/local-storage": "^6.0.0-rc.0",
+    "angular2-image-upload": "^1.0.0-rc.1",
     "chai": "^4.1.2",
-    "core-js": "^2.4.1",
+    "core-js": "^2.5.7",
     "crypto": "^1.0.1",
-    "mocha": "^5.0.0",
-    "ng-recaptcha": "^3.0.3",
+    "mocha": "^5.2.0",
+    "ng-recaptcha": "^3.0.5",
     "ng2-cookies": "^1.0.12",
-    "ng2-toastr": "^4.1.2",
     "ngx-auto-scroll": "^1.1.0",
+    "ngx-toastr": "^8.8.0",
     "node-datetime": "^2.0.6",
     "node.date-time": "^1.2.2",
-    "nodemailer": "^4.4.2",
-    "rxjs": "^5.5.6",
+    "nodemailer": "^4.6.7",
+    "rxjs": "^6.2.1",
+    "rxjs-compat": "^6.2.1",
     "should": "^13.2.1",
-    "socket.io": "^2.1.0",
-    "zone.js": "^0.8.19"
+    "socket.io": "^2.1.1",
+    "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "1.6.6",
-    "@angular/compiler-cli": "^5.2.0",
-    "@angular/language-service": "^5.2.0",
-    "@types/jasmine": "~2.8.3",
+    "@angular/cli": "^6.0.8",
+    "@angular/compiler-cli": "^6.0.5",
+    "@angular/language-service": "^6.0.5",
+    "@types/jasmine": "^2.8.8",
     "@types/jasminewd2": "~2.0.2",
-    "@types/node": "~6.0.60",
-    "codelyzer": "^4.0.1",
+    "@types/node": "^6.0.113",
+    "codelyzer": "^4.3.0",
     "jasmine-core": "~2.8.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~2.0.0",
+    "karma": "^2.0.3",
     "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "^1.2.1",
-    "karma-jasmine": "~1.1.0",
+    "karma-coverage-istanbul-reporter": "^1.4.3",
+    "karma-jasmine": "^1.1.2",
     "karma-jasmine-html-reporter": "^0.2.2",
     "protractor": "~5.1.2",
     "ts-node": "~4.1.0",
     "tslint": "~5.9.1",
-    "typescript": "~2.5.3"
+    "typescript": "~2.7.2",
+    "@angular-devkit/build-angular": "~0.6.8"
   }
 }
