@@ -1,7 +1,6 @@
 import {Component, OnInit, ViewContainerRef} from '@angular/core';
-import {FormControl, FormGroup} from "@angular/forms";
-import {ForgotService} from "./forgot.service";
-import { ToastsManager } from 'ng2-toastr/ng2-toastr';
+import {FormControl, FormGroup} from '@angular/forms';
+import {ForgotService} from '../../services/forgot.service';
 
 @Component({
   selector: 'app-forgot',
@@ -12,21 +11,20 @@ export class ForgotComponent implements OnInit {
 
   private forgotService;
 
-  constructor(forgotService: ForgotService,public toastr: ToastsManager, vcr: ViewContainerRef) {
+  constructor(forgotService: ForgotService) {
     this.forgotService = forgotService;
-    this.toastr.setRootViewContainerRef(vcr);
 
   }
 
 
   retypeEmailFrom = new FormGroup ({
     email: new FormControl(),
-  })
+  });
 
 
 
-  RetypeEmail(){
-    this.forgotService.forgotPassword(this.retypeEmailFrom.controls.email.value)
+  RetypeEmail() {
+    this.forgotService.forgotPassword(this.retypeEmailFrom.controls.email.value);
 
   }
 
