@@ -1,21 +1,20 @@
 import {Injectable} from '@angular/core';
-import {HttpClient} from "@angular/common/http";
-import { ToastsManager } from 'ng2-toastr/ng2-toastr';
-import {Router} from "@angular/router";
-import Config from "../../../../app-config";
+import {HttpClient} from '@angular/common/http';
+import {Router} from '@angular/router';
+import Config from '../../../../app-config';
 
 @Injectable()
 export class SignupService {
 
   private testUrl = Config.nodeApi;
-  constructor( private http: HttpClient, private push: ToastsManager, private router:Router) {}
+  constructor( private http: HttpClient, private router: Router) {}
 
-  signUserIn (User,captcha) {
+  signUserIn (User, captcha, number?: number) {
 
 
     this.http
-      .post(this.testUrl +'users/signup', {companyName: User.companyName,email: User.email,
-        password: User.password, image: User.image,captcha:captcha}, )
+      .post(this.testUrl + 'users/signup', {companyName: User.companyName, email: User.email,
+        password: User.password, image: User.image, captcha: captcha, phone: number}, )
       .subscribe(
         // Successful responses call the first callback.
         data => {
@@ -24,7 +23,7 @@ export class SignupService {
         },
         // Errors will call this callback instead:
         err => {
-          this.push.error('Something went wrong');
+          // this.push.error('Something went wrong');
         }
       );}
     saveImage(img) {
