@@ -1,53 +1,50 @@
 import {Injectable} from '@angular/core';
-import {HttpClient} from "@angular/common/http";
-import {ToastsManager } from 'ng2-toastr/ng2-toastr';
-import {Router} from "@angular/router";
-import {UserService} from "./user.service";
-import Config from "../../../app-config";
-import { Cookie } from 'ng2-cookies/ng2-cookies';
+import {HttpClient} from '@angular/common/http';
+import { ToastrService } from 'ngx-toastr';
+import {Router} from '@angular/router';
+import {UserService} from './user.service';
+import Config from '../../../app-config';
+import { LocalStorage } from '@ngx-pwa/local-storage';
 
 @Injectable()
 export class LoginService {
 
-   message;
-  private testUrl = Config.nodeApi;
+  private url = Config.nodeApi;
 
-  constructor( private http: HttpClient, private push: ToastsManager, private router:Router,
-               private userService: UserService) {
+  constructor( private http: HttpClient, private push: ToastrService, private router: Router,
+               private userService: UserService, protected localStorage: LocalStorage) {
   }
 
   logUserIn (email, pass) {
 
-    return new Promise((resolve, reject) =>{
-      this.http
-        .post(this.testUrl +'users/login', {email: email, password: pass}, )
-        .subscribe(
-          // Successful responses call the first callback.
-          data => {
-            console.log('success login');
-
-            var message = data['message'];
-            resolve(message);
-            if(data['success']) {
-              let token = data['token'];
-              let user = data['user'];
-              Cookie.set('token', token );
+    return new Promise((resolve, reject) => {
+      this.http.post(this.url + 'users/login', {email: email, password: pass})
+        .subscribe(data => {
+            const message = data['message'];
+            if (data['success']) {
+              const token = data['token'];
               this.router.navigate(['']);
-              // this.userService.setUser(data['user']);
-              resolve(data['message']);
+
+              this.localStorage.setItem('token', data['token']).subscribe((val) => {
+
+                this.userService.setUser(data['user'], data['token']);
+                resolve(data['message']);
+              });
             }
+            resolve(message);
           },
           error => { // Error
             reject(error);
           }
         );
-    })
+    });
   }
 
-  logout(){
-    Cookie.delete("token");
-    // this.router.navigate(['login']);
-    this.userService.setUser(null);
+  logout() {
+    this.localStorage.removeItem('token').subscribe((val) => {
+      this.userService.setUser(null);
+      this.push.success('logout completed');
+    });
   }
 
 }
